+++ src/module.js
@@ -43,6 +43,13 @@
  * the child state.  However, the controller of the abstract state will be
  * initialized AFTER any configuration rules of the child state have been met.
  *
+ * # Support for `data.authorities`
+ *
+ * If you have used [JHipster](https://jhipster.github.io/) to generate your
+ * project, you are likely using the `data.authorities` property to define
+ * authorization for your views. This library will look for the `data.authorities`
+ * property and apply the same logic as our own `sp.authorize` property.
+ *
  * @example
  *
  * <pre>
@@ -82,6 +89,20 @@
  *       });
  * });
  * </pre>
+ *
+ * If using JHipster generated code:
+ *
+ *  <pre>
+ *     // Require a user to be in the admins group in order to see this state
+ *     $stateProvider
+ *       .state('secrets', {
+ *         url: '/admin',
+ *         controller: 'AdminCtrl',
+ *         data: {
+ *           authorities: ['admins']
+ *         }
+ *       });
+ *  </pre>
  */
 
  /**
@@ -332,13 +353,14 @@ angular.module('stormpath', [
       StormpathService.prototype.stateChangeInterceptor = function stateChangeInterceptor(config) {
         $rootScope.$on('$stateChangeStart', function(e,toState,toParams){
           var sp = toState.sp || {}; // Grab the sp config for this state
+          var authorities = (toState.data && toState.data.authorities) ? toState.data.authorities : undefined;
 
-          if((sp.authenticate || sp.authorize) && (!$user.currentUser)){
+          if((sp.authenticate || sp.authorize || (authorities && authorities.length)) && (!$user.currentUser)){
             e.preventDefault();
             $user.get().then(function(){
               // The user is authenticated, continue to the requested state
-              if(sp.authorize){
-                if(authorizeStateConfig(sp)){
+              if(sp.authorize || (authorities && authorities.length)){
+                if(authorizeStateConfig(sp, authorities)){
                   $state.go(toState.name,toParams);
                 }else{
                   stateChangeUnauthorizedEvent(toState,toParams);
@@ -356,8 +378,8 @@ angular.module('stormpath', [
               $state.go(toState.name,toParams);
             });
           }
-          else if($user.currentUser && sp.authorize){
-            if(!authorizeStateConfig(sp)){
+          else if($user.currentUser && (sp.authorize || (authorities && authorities.length))){
+            if(!authorizeStateConfig(sp, authorities)){
               e.preventDefault();
               stateChangeUnauthorizedEvent(toState,toParams);
             }
@@ -381,15 +403,21 @@ angular.module('stormpath', [
         });
       };
 
-      function authorizeStateConfig(spStateConfig){
+      function authorizeStateConfig(spStateConfig, authorities){
         var sp = spStateConfig;
         if(sp && sp.authorize && sp.authorize.group) {
           return $user.currentUser.inGroup(sp.authorize.group);
+        }else if(authorities){
+          // add support for reading from JHipster's data: { authorities: ['ROLE_ADMIN'] }
+          // https://github.com/stormpath/stormpath-sdk-angularjs/issues/190
+          var roles = authorities.filter(function(authority){
+            return $user.currentUser.inGroup(authority);
+          });
+          return roles.length > 0;
         }else{
           console.error('Unknown authorize configuration for spStateConfig',spStateConfig);
           return false;
         }
-
       }
 
       function routeChangeUnauthenticatedEvent(toRoute) {