+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.1.1
+
+Our UI Router integration will now look for `data.authorities` on UI state configuration, and apply the same behavior as `sp.authorize.group`.  This is a convenience for [JHipster](https://jhipster.github.io/) users, as they
+already use `data.authorities` to declare the roles that a user must have in order to access a view.
+
 # 1.1.0
 
 * The `$user.get()` method now takes a boolean `bypassCache` option, e.g. `$user.get(true)` will make a new request of the `/me` endpoint, instead of returning the user data that was already cached on login.